11.28yaya项目

+++ app/src/main/java/com/administrator/yaya/bean/homepage/TestHomePageData.java
@@ -126,7 +126,7 @@ public class TestHomePageData implements Serializable {
             private int userStatus;
             private String userInvitationCode;
             private int userParentId;
-            private int userEarningsNow;
+            private double userEarningsNow;
             private int userEarningsTotal;
             private int userContributeTotal;
             private int zfbEd;
@@ -279,11 +279,11 @@ public class TestHomePageData implements Serializable {
                 this.userParentId = userParentId;
             }
 
-            public int getUserEarningsNow() {
+            public double getUserEarningsNow() {
                 return userEarningsNow;
             }
 
-            public void setUserEarningsNow(int userEarningsNow) {
+            public void setUserEarningsNow(double userEarningsNow) {
                 this.userEarningsNow = userEarningsNow;
             }
