+++ app/src/main/java/com/administrator/yaya/bean/my/TestIncome.java
@@ -99,7 +99,7 @@ public class TestIncome {
             private int userStatus;
             private String userInvitationCode;
             private int userParentId;
-            private int userEarningsNow;
+            private double userEarningsNow;
             private int userEarningsTotal;
             private int userContributeTotal;
             private int zfbEd;
@@ -246,11 +246,11 @@ public class TestIncome {
                 this.userParentId = userParentId;
             }
 
-            public int getUserEarningsNow() {
+            public double getUserEarningsNow() {
                 return userEarningsNow;
             }
 
-            public void setUserEarningsNow(int userEarningsNow) {
+            public void setUserEarningsNow(double userEarningsNow) {
                 this.userEarningsNow = userEarningsNow;
             }
