+++ app/src/main/java/com/administrator/yaya/bean/my/TestMyEarnings.java
@@ -121,7 +121,7 @@ public class TestMyEarnings {
             private int userStatus;
             private String userInvitationCode;
             private int userParentId;
-            private int userEarningsNow;
+            private double userEarningsNow;
             private int userEarningsTotal;
             private int userContributeTotal;
             private int zfbEd;
@@ -268,11 +268,11 @@ public class TestMyEarnings {
                 this.userParentId = userParentId;
             }
 
-            public int getUserEarningsNow() {
+            public double getUserEarningsNow() {
                 return userEarningsNow;
             }
 
-            public void setUserEarningsNow(int userEarningsNow) {
+            public void setUserEarningsNow(double userEarningsNow) {
                 this.userEarningsNow = userEarningsNow;
             }
